fix(prefer-tacit): only apply fixer for TS >= 4.7
