+++ package.json
@@ -64,7 +64,8 @@
   "dependencies": {
     "@typescript-eslint/utils": "^5.10.2",
     "deepmerge-ts": "^4.0.3",
-    "escape-string-regexp": "^4.0.0"
+    "escape-string-regexp": "^4.0.0",
+    "semver": "^7.3.7"
   },
   "devDependencies": {
     "@commitlint/cli": "^16.1.0",
@@ -87,6 +88,7 @@
     "@types/estree": "^0.0.51",
     "@types/node": "18.0.0",
     "@types/rollup-plugin-auto-external": "^2.0.2",
+    "@types/semver": "^7.3.12",
     "@typescript-eslint/eslint-plugin": "^5.10.2",
     "@typescript-eslint/parser": "^5.10.2",
     "ava": "^4.0.1",
