+++ tests/rules/prefer-tacit/ts/invalid.ts
@@ -148,44 +148,6 @@ const tests: ReadonlyArray<InvalidTestCase> = [
   {
     code: dedent`
       function f<T>(x: T): T {}
-      const foo = x => f<number>(x);
-    `,
-    optionsSet: [[]],
-    output: dedent`
-      function f<T>(x: T): T {}
-      const foo = f<number>;
-    `,
-    errors: [
-      {
-        messageId: "generic",
-        type: "ArrowFunctionExpression",
-        line: 2,
-        column: 13,
-      },
-    ],
-  },
-  {
-    code: dedent`
-      function f<T>(x: T): T {}
-      function foo(x) { return f<number>(x); }
-    `,
-    optionsSet: [[]],
-    output: dedent`
-      function f<T>(x: T): T {}
-      const foo = f<number>;
-    `,
-    errors: [
-      {
-        messageId: "generic",
-        type: "FunctionDeclaration",
-        line: 2,
-        column: 1,
-      },
-    ],
-  },
-  {
-    code: dedent`
-      function f<T>(x: T): T {}
       export default function (x) { return f<number>(x); }
     `,
     optionsSet: [[]],
