+++ tests/rules/prefer-tacit/ts4.7/index.ts
@@ -1,0 +1,7 @@
+import invalid from "./invalid";
+import valid from "./valid";
+
+export default {
+  valid,
+  invalid,
+};
