+++ tests/rules/prefer-tacit/ts4.7/valid.ts
@@ -1,0 +1,15 @@
+import dedent from "dedent";
+
+import type { ValidTestCase } from "~/tests/helpers/util";
+
+const tests: ReadonlyArray<ValidTestCase> = [
+  // Instantiation Expression
+  {
+    code: dedent`
+      const foo = f<number>;
+    `,
+    optionsSet: [[]],
+  },
+];
+
+export default tests;
