+++ tests/rules/prefer-tacit/tsNot4.7/valid.ts
@@ -1,0 +1,7 @@
+import dedent from "dedent";
+
+import type { ValidTestCase } from "~/tests/helpers/util";
+
+const tests: ReadonlyArray<ValidTestCase> = [];
+
+export default tests;
